requirements.txt
click==8.1.7
Faker==25.8.0
grpcio==1.63.0
grpcio-tools==1.63.0
overrides==7.7.0
protobuf==5.26.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
robotframework==7.0.1
robotframework-assertion-engine==3.0.3
robotframework-browser==18.5.1
robotframework-pythonlibcore==4.4.1
six==1.16.0
typing_extensions==4.12.2
wrapt==1.16.0
